Adds folder Constants and Types
