+++ src/Constants/Constants.tsx
@@ -1,0 +1,7 @@
+import { showDate } from '../utils/showDate';
+
+export const BASE_URL = 'http://api.openweathermap.org/data/2.5/';
+export const API_KEY = '57a90a8fcac701e57c9c008de996f569';
+export const DEFAULT_CITY = 'Minsk';
+export const cities = ['Minsk', 'Moscow', 'Bratislava', 'Kiev'];
+export const date: string = showDate();
