+++ src/components/CurrentInfo/CurrentInfo.tsx
@@ -1,14 +1,8 @@
 import { ButtonsList } from '../ButtonsList/ButtonsList';
 import { CardInfo } from '../CurrentInfo/CardInfo';
 import { Title } from '../Title/Title';
-import { IWeather } from '../../pages/CurrentWeather/CurrentWeather';
+import { CurrentIngoProps } from '../../types/types';
 import style from '../CurrentInfo/CurrentInfo.module.scss';
-
-type CurrentIngoProps = {
-  weather: IWeather;
-  setCity: (city: string) => void;
-  city: string;
-};
 
 export const CurrentInfo = ({ weather, setCity, city }: CurrentIngoProps) => {
   return (
