+++ src/components/Notification/Notification.tsx
@@ -1,8 +1,5 @@
+import { NotificationProps } from '../../types/types';
 import style from '../Notification/Notification.module.scss';
-
-type NotificationProps = {
-  message: string;
-};
 
 export const Notification = ({ message }: NotificationProps) => (
   <h1 className={style.notificationError}>{message}</h1>
