+++ src/components/Title/Title.tsx
@@ -1,8 +1,5 @@
+import { TitleProps } from '../../types/types';
 import style from '../Title/Title.module.scss';
-
-type TitleProps = {
-  title: string;
-};
 
 export const Title = ({ title }: TitleProps) => (
   <h1 className={style.currentWeatherTitle}>{title}</h1>
